Cleaner Code and Comments added.
Cleaned the code for the calculator (e.g. Fixed Spelling mistakes and Adjusted spacing) and added comments outlining each part of the code. Percentage and Bracts are not operational for the security reasons but soon will be. Updates regarding the Security, Responsiveness, and the Functionality of the program will be coming soon :).